+++ docs/examples/basic_context/child_context.py
@@ -11,3 +11,19 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from opentelemetry import baggage, trace
+
+tracer = trace.get_tracer(__name__)
+
+global_ctx = baggage.set_baggage("context", "global")
+with tracer.start_as_current_span(name="root span") as root_span:
+    parent_ctx = baggage.set_baggage("context", "parent")
+    with tracer.start_as_current_span(
+        name="child span", context=parent_ctx
+    ) as child_span:
+        child_ctx = baggage.set_baggage("context", "child")
+
+print(baggage.get_baggage("context", global_ctx))
+print(baggage.get_baggage("context", parent_ctx))
+print(baggage.get_baggage("context", child_ctx))