+++ exporter/opentelemetry-exporter-opencensus/tests/test_otcollector_trace_exporter.py
@@ -96,7 +96,7 @@ class TestCollectorSpanExporter(unittest.TestCase):
             trace_flags=TraceFlags(TraceFlags.SAMPLED),
             trace_state=trace_api.TraceState({"testKey": "testValue"}),
         )
-        parent_context = trace_api.SpanContext(
+        parent_span_context = trace_api.SpanContext(
             trace_id, parent_id, is_remote=False
         )
         other_context = trace_api.SpanContext(
@@ -118,27 +118,27 @@ class TestCollectorSpanExporter(unittest.TestCase):
             context=other_context, attributes=link_attributes
         )
         link_2 = trace_api.Link(
-            context=parent_context, attributes=link_attributes
+            context=parent_span_context, attributes=link_attributes
         )
-        span_1 = trace.Span(
+        span_1 = trace._Span(
             name="test1",
             context=span_context,
-            parent=parent_context,
+            parent=parent_span_context,
             events=(event,),
             links=(link_1,),
             kind=trace_api.SpanKind.CLIENT,
         )
-        span_2 = trace.Span(
+        span_2 = trace._Span(
             name="test2",
-            context=parent_context,
+            context=parent_span_context,
             parent=None,
             kind=trace_api.SpanKind.SERVER,
         )
-        span_3 = trace.Span(
+        span_3 = trace._Span(
             name="test3",
             context=other_context,
             links=(link_2,),
-            parent=span_2.get_context(),
+            parent=span_2.get_span_context(),
         )
         otel_spans = [span_1, span_2, span_3]
         otel_spans[0].start(start_time=start_times[0])
@@ -302,7 +302,7 @@ class TestCollectorSpanExporter(unittest.TestCase):
             trace_flags=TraceFlags(TraceFlags.SAMPLED),
         )
         otel_spans = [
-            trace.Span(
+            trace._Span(
                 name="test1",
                 context=span_context,
                 kind=trace_api.SpanKind.CLIENT,