+++ instrumentation/opentelemetry-instrumentation-celery/tests/test_utils.py
@@ -42,7 +42,7 @@ class TestUtils(unittest.TestCase):
             "routing_key": "celery",
         }
 
-        span = trace.Span("name", mock.Mock(spec=trace_api.SpanContext))
+        span = trace._Span("name", mock.Mock(spec=trace_api.SpanContext))
         utils.set_attributes_from_context(span, context)
 
         self.assertEqual(
@@ -69,6 +69,30 @@ class TestUtils(unittest.TestCase):
         )
         self.assertNotIn("custom_meta", span.attributes)
 
+    def test_set_attributes_not_recording(self):
+        # it should extract only relevant keys
+        context = {
+            "correlation_id": "44b7f305",
+            "delivery_info": {"eager": True},
+            "eta": "soon",
+            "expires": "later",
+            "hostname": "localhost",
+            "id": "44b7f305",
+            "reply_to": "44b7f305",
+            "retries": 4,
+            "timelimit": ("now", "later"),
+            "custom_meta": "custom_value",
+            "routing_key": "celery",
+        }
+
+        mock_span = mock.Mock()
+        mock_span.is_recording.return_value = False
+        utils.set_attributes_from_context(mock_span, context)
+        self.assertFalse(mock_span.is_recording())
+        self.assertTrue(mock_span.is_recording.called)
+        self.assertFalse(mock_span.set_attribute.called)
+        self.assertFalse(mock_span.set_status.called)
+
     def test_set_attributes_from_context_empty_keys(self):
         # it should not extract empty keys
         context = {
@@ -78,7 +102,7 @@ class TestUtils(unittest.TestCase):
             "retries": 0,
         }
 
-        span = trace.Span("name", mock.Mock(spec=trace_api.SpanContext))
+        span = trace._Span("name", mock.Mock(spec=trace_api.SpanContext))
         utils.set_attributes_from_context(span, context)
 
         self.assertEqual(len(span.attributes), 0)
@@ -99,7 +123,7 @@ class TestUtils(unittest.TestCase):
 
         # propagate and retrieve a Span
         task_id = "7c6731af-9533-40c3-83a9-25b58f0d837f"
-        span = trace.Span("name", mock.Mock(spec=trace_api.SpanContext))
+        span = trace._Span("name", mock.Mock(spec=trace_api.SpanContext))
         utils.attach_span(fn_task, task_id, span)
         span_after = utils.retrieve_span(fn_task, task_id)
         self.assertIs(span, span_after)
@@ -112,7 +136,7 @@ class TestUtils(unittest.TestCase):
 
         # propagate a Span
         task_id = "7c6731af-9533-40c3-83a9-25b58f0d837f"
-        span = trace.Span("name", mock.Mock(spec=trace_api.SpanContext))
+        span = trace._Span("name", mock.Mock(spec=trace_api.SpanContext))
         utils.attach_span(fn_task, task_id, span)
         # delete the Span
         utils.detach_span(fn_task, task_id)