+++ instrumentation/opentelemetry-instrumentation-falcon/src/opentelemetry/instrumentation/falcon/__init__.py
@@ -53,7 +53,10 @@ from opentelemetry import configuration, context, propagators, trace
 from opentelemetry.configuration import Configuration
 from opentelemetry.instrumentation.falcon.version import __version__
 from opentelemetry.instrumentation.instrumentor import BaseInstrumentor
-from opentelemetry.instrumentation.utils import http_status_to_canonical_code
+from opentelemetry.instrumentation.utils import (
+    extract_attributes_from_object,
+    http_status_to_canonical_code,
+)
 from opentelemetry.trace.status import Status
 from opentelemetry.util import ExcludeList, time_ns
 
@@ -114,13 +117,16 @@ class _InstrumentedFalconAPI(falcon.API):
         token = context.attach(
             propagators.extract(otel_wsgi.get_header_from_environ, env)
         )
-        attributes = otel_wsgi.collect_request_attributes(env)
         span = self._tracer.start_span(
             otel_wsgi.get_default_span_name(env),
             kind=trace.SpanKind.SERVER,
-            attributes=attributes,
             start_time=start_time,
         )
+        if span.is_recording():
+            attributes = otel_wsgi.collect_request_attributes(env)
+            for key, value in attributes.items():
+                span.set_attribute(key, value)
+
         activation = self._tracer.use_span(span, end_on_exit=True)
         activation.__enter__()
         env[_ENVIRON_SPAN_KEY] = span
@@ -159,17 +165,18 @@ class _TraceMiddleware:
 
     def process_request(self, req, resp):
         span = req.env.get(_ENVIRON_SPAN_KEY)
-        if not span:
+        if not span or not span.is_recording():
             return
 
-        for attr in self._traced_request_attrs:
-            value = getattr(req, attr, None)
-            if value is not None:
-                span.set_attribute(attr, str(value))
+        attributes = extract_attributes_from_object(
+            req, self._traced_request_attrs
+        )
+        for key, value in attributes.items():
+            span.set_attribute(key, value)
 
     def process_resource(self, req, resp, resource, params):
         span = req.env.get(_ENVIRON_SPAN_KEY)
-        if not span:
+        if not span or not span.is_recording():
             return
 
         resource_name = resource.__class__.__name__
@@ -182,7 +189,7 @@ class _TraceMiddleware:
         self, req, resp, resource, req_succeeded=None
     ):  # pylint:disable=R0201
         span = req.env.get(_ENVIRON_SPAN_KEY)
-        if not span:
+        if not span or not span.is_recording():
             return
 
         status = resp.status